Initial commit for version tracking
Here I added all the files to a new git repository for tracking. The code
is incomplete in this commit and should not be built or trusted in any way.
This really is basically just copy-paste from the OSDev wiki to get some
base code and isn't its own operating system yet.

Signed-off-by: Nkeck720 <noahkeck72@gmail.com>